Move Core Publics so imports match between SPM and CocoaPods
