+++ FirebaseCore/Sources/FIRCoreDiagnosticsConnector.m
@@ -18,7 +18,7 @@
 
 #import "Interop/CoreDiagnostics/Public/FIRCoreDiagnosticsInterop.h"
 
-#import "FirebaseCore/Sources/Public/FIROptions.h"
+#import "FirebaseCore/Sources/Public/FirebaseCore/FIROptions.h"
 
 #import "FirebaseCore/Sources/FIRDiagnosticsData.h"
 #import "FirebaseCore/Sources/Private/FIRAppInternal.h"
