+++ FirebaseCore/Sources/FIRLogger.m
@@ -14,7 +14,7 @@
 
 #import "FirebaseCore/Sources/Private/FIRLogger.h"
 
-#import "FirebaseCore/Sources/Public/FIRLoggerLevel.h"
+#import "FirebaseCore/Sources/Public/FirebaseCore/FIRLoggerLevel.h"
 #import "GoogleUtilities/Environment/Private/GULAppEnvironmentUtil.h"
 #import "GoogleUtilities/Logger/Private/GULLogger.h"
