+++ FirebaseCore/Sources/Private/FIRLogger.h
@@ -16,12 +16,7 @@
 
 #import <Foundation/Foundation.h>
 
-#if SWIFT_PACKAGE
-// TODO(paulb777): Investigate if there's a common strategy for both Cocoapods and Swift PM.
-#import "FIRLoggerLevel.h"
-#else
 #import <FirebaseCore/FIRLoggerLevel.h>
-#endif
 
 NS_ASSUME_NONNULL_BEGIN
