+++ FirebaseCore/Sources/Private/FIROptionsInternal.h
@@ -14,12 +14,7 @@
  * limitations under the License.
  */
 
-#if SWIFT_PACKAGE
-// TODO(paulb777): Investigate if there's a common strategy for both Cocoapods and Swift PM.
-#import "FIROptions.h"
-#else
 #import <FirebaseCore/FIROptions.h>
-#endif
 
 /**
  * Keys for the strings in the plist file.
