+++ FirebaseCore/Sources/Private/FirebaseCoreInternal.h
@@ -15,11 +15,7 @@
 // An umbrella header, for any other libraries in this repo to access Firebase Public and Private
 // headers. Any package manager complexity should be handled here.
 
-#if SWIFT_PACKAGE
-@import FirebaseCore;
-#else
 #import <FirebaseCore/FirebaseCore.h>
-#endif
 
 #import "FirebaseCore/Sources/Private/FIRAppInternal.h"
 #import "FirebaseCore/Sources/Private/FIRComponent.h"
