+++ FirebaseCore/Sources/Public/FirebaseCore/FIRConfiguration.h
